requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
arrow==1.4.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.2
bleach[css]==6.2.0
boto3==1.40.58
botocore==1.40.58
cairocffi==1.7.1
cairosvg==2.8.2
cbor2==5.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.11.0
cryptography==46.0.3
cssselect2==0.8.0
curl-cffi==0.13.0
defusedxml==0.7.1
diskcache==5.6.3
django==5.2.8
django-admincommand @ git+https://github.com/rphlo/django-admincommand.git
django-allauth==65.12.1
django-appconf==1.1.0
django-background-tasks @ git+https://github.com/rphlo/django-background-tasks@6b247d0b3a7565613656d0ec07b55f6c66aa03cc
django-bootstrap5 @ git+https://github.com/rphlo/django-bootstrap5.git
django-compressor==4.5.1
django-cors-headers==4.9.0
django-csp==4.0
django-environ==0.12.0
django-health-check==3.20.0
django-hijack==3.7.4
django-hosts==7.0.0
django-markdownify==0.9.5
django-minify-html==1.13.0
django-oauth-toolkit==3.1.0
django-permissions-policy==4.28.0
django-qr-code==4.2.0
django-s3-storage==0.15.0
django-user-sessions @ git+https://github.com/jazzband/django-user-sessions@1ad42f36758251f221e976683eca91a8c79df09e
djangorestframework==3.16.1
drf-orjson-renderer==1.7.3
drf-yasg==1.21.11
emoji-country-flag==2.1.0
frozenlist==1.8.0
geoip2==5.1.0
geojson-validator==0.6.0
gps-data-codec @ git+https://github.com/routechoiceslivegps/gps-data-codec
gpxpy==1.6.2
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jmespath==1.0.1
jwcrypto==1.5.6
kagi @ git+https://github.com/rphlo/kagi@9d02f068abf0963575819455f2e3f05888726153
loguru==0.7.3
lxml==6.0.2
markdown==3.9
maxminddb==2.8.2
minify-html==0.18.0
multidict==6.7.0
numpy==2.3.4
oauthlib==3.3.1
opencv-python==4.11.0.86
orjson==3.11.4
packaging==25.0
pillow==12.0.0
pillow-jxl-plugin==1.3.5
propcache==0.4.1
psycopg[binary,pool]==3.2.11
psycopg-binary==3.2.11
psycopg-pool==3.2.6
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pyopenssl==25.3.0
pypng==0.20220715.0
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.2
pyyaml==6.0.3
qrcode==7.4.2
rcssmin==1.1.2
requests==2.32.5
reverse-geocode==1.6.6
rjsmin==1.2.2
s3transfer==0.14.0
segno==1.6.6
sentry-sdk==2.42.1
sewer @ git+https://github.com/rphlo/sewer@0.8.4b
shapely==2.1.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
timezonefinder==8.1.0
tinycss2==1.4.0
tornado==6.5.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webauthn==1.11.1
webencodings==0.5.1
yarl==1.22.0
